Move feature and class parsing to top level
Feature and Class parsing seperated from main character body
